Aspectos de qualidade

+++ BackECommerceTest/CarrinhoTests.cs
@@ -218,12 +218,14 @@ namespace BackECommerceTest
             Produto prod = _produtoRepository.BuscarProduto("5ebd72407f288a429408e5f9");
 
             Carrinho carrinhoResult = _carrinhoRepository.AddProduto(_usuarioTeste.Id, prod.Id);
-            //if (carrinhoResult != null)
-            //{
-            //    Assert.Fail();
-            //}
-
-            _carrinhoRepository.RemoverCarrinhoPorUsuario(_usuarioTeste.Id);
+            if (carrinhoResult != null)
+            {
+                Assert.Fail();
+            }
+            else
+            {
+                _carrinhoRepository.RemoverCarrinhoPorUsuario(_usuarioTeste.Id);
+            }            
         }
 
         [Test]
